admin_ui/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyjwt>=2.9.0
bcrypt>=4.2.0
